requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
filelock==3.9.0
Markdown==3.4.1
Pillow==9.4.0
platformdirs==2.6.2
psycopg2==2.9.5
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
virtualenv==20.17.1
